learning_users/requirements.txt
argon2-cffi==19.1.0
bcrypt==3.1.7
cffi==1.12.3
Django==2.2.5
django-filter==2.2.0
djangorestframework==3.10.3
Faker==2.0.2
Pillow==6.1.0
pycparser==2.19
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.3
mysqlclient==1.4.4
